Create SafeUser type
In older versions of Next, you got a warning when trying to pass a User object
with values of type `Date` from a server component to client components.

In this commit, we update getCurrentUser to return a currentUser object where
properties of type `Date` are converted to a `string`.

We also create a SafeUser type, which is based on User but changes properties
of type `Date` to `string`. We then update the places where we use
`currentUser` to use the new `SafeUser` type