Add missing account.type column
This was causing OAuth via GitHub to not work correctly.

The user was being added to the users table, but the insertion into the
accounts table was failing due to a missing type column. Apparently, this
insertion is necessary for NextAuth to create a session when logging in via
OAuth. So the user was being created, but the user was NOT being signed in.