Update links in Register/ Login modal footers
When clicking the link in the Register modal footer, it should close the
Register modal and open the Login modal. The reverse should happen in the Login
modal.

Update the copy of the link in the Login modal and update the onClick callback
in both modals to achieve this functionality.