Implement SearchModal
Clicking on the Search component in the NavBar should open a modal where a user
can select:
- location
- dates
- guest, room, bathroom count

Clicking the "Search" button on the last pane should close the modal and update
the URL with the new listing search criteria.

NEXT UP: udate the `getListings` action to use the URL's query params to
filter listings based on what the user is looking for.