Use prisma User type instead of SafeUser
The SafeUser type converted certain fields on the User returned by Prisma to
strings instead of date objects. This seems to be unnecessary and was leading
to some TypeScript errors because in certain places I was using a User and
others a SafeUser. The easiest fix was to remove SafeUser altogether and just
use the User returned from the database.

In this commit, I remove the SafeUser type and update each place this type was
used to expect a User type instead.